Add better error handling
This changes the return signature for the Value() method to also return
an error. Returning an error in this context provides an opportunity to
continue parsing the rest of the file while skipping the problem record.
The error that might be returned from the Err() method would typically
indicate something deeply wrong with the input file that would likely
need to be addressed before continuing.