Improve parser performance
There are two main changes in this commit. The first changes to storing
the fields in a hash rather than a slice. This makes pulling a specific
field from the record faster. The second change delays parsing a field
into its struct until necessary. The initial pass simply stores the byte
array of the field data in the hash, keyed to the tag.

Overall performance improvement is about 50%.