Reduce the number of heap allocations during parse
This reduces the number of heap allocations when scanning a record into
a Record struct. The total amount of heap requested probably won't
change much, but instead of doing it piecemeal it does it once for each
record. The overall performance improvement is about 30%.

This is almost an entirely internal optimization. The only aspect of the
public interface that changes is that a Record's Data field changes from
a byte slice to a string. I don't think mario is using this.